Moved Block and Pipeline components to partials so they can be reused.
