+++ src/components/Partials/Block.vue
@@ -1,42 +1,37 @@
-<<template>
+<template>
 <div>
-
     <div v-if="!block">
-          <b-row>
-              </b-row>
-              </div>
-    <div v-else class="block">
         <b-row>
-            <b-col>
+        </b-row>
+    </div>
+    <div v-else class="block">
         <span class="namespace">{{ block.Namespace }}</span>
-        <h3>{{ block.Name }}</h3>
+        <h3>
+            <router-link :to="{ name: 'blocks', params: { blockname: `${block.Namespace}.${block.Name}` }}">{{ block.Name }}</router-link>
+        </h3>
 
-        <div v-bind:title="block.Receives" class="code">
+        <div v-bind:title="'Input: '+block.Receives" class="code">
             <i class="fas fa-sign-in-alt"></i> {{ block.Receives | prettyClrType }}
         </div>
-        <div v-bind:title="block.Returns" class="code">
+        <div v-bind:title="'Output: '+block.Returns" class="code">
             <i class="fas fa-sign-out-alt"></i> {{ block.Returns | prettyClrType }}
         </div>
-        </b-col>
-        </b-row>
-        <b-row class="mt-3">
-            <b-col>
-        <h4>Used in the following pipelines:</h4>
-        <ul>
-            <li v-for="pipeline in pipelines">
-                <router-link :to="{ name: 'pipelines', params: { pipelineid: getPipelineName(pipeline) }}">
-                {{ pipeline.Name }}
-                </router-link>
-            </li>
-        </ul>
-            </b-col>
-        </b-row>
+        <div>
+            <h4>Used in the following pipelines:</h4>
+            <ul>
+                <li v-for="pipeline in pipelines">
+                    <router-link :to="{ name: 'pipelines', params: { pipelineid: getPipelineName(pipeline) }}">
+                    {{ pipeline.Name }}
+                    </router-link>
+                </li>
+            </ul>
+        </div>
     </div>
 </div>
 </template>
 
 <script>
-import { prettyClrType } from "../filters/clrTypes";
+import { prettyClrType } from "../../filters/clrTypes";
 
 export default {
   props: ["blockname"],
@@ -67,6 +62,12 @@ export default {
   methods: {
     getPipelineName: pipeline => {
       return `${pipeline.Namespace}.${pipeline.Name}`;
+    },
+    getPipeline: function (namespace, name) {
+      var pipeline = this.pipelines.find(pipeline => {
+        return `${pipeline.Namespace}.${pipeline.Name}` == `${namespace}.${name}`;
+      });
+      return pipeline;
     }
   }
 };
@@ -74,10 +75,13 @@ export default {
 
 <style>
 .block {
-  background-color: #e1bee7 !important;
+  background-color: #aed581 !important;
   padding: 10px;
   -webkit-border-radius: 8px;
   -moz-border-radius: 8px;
   border-radius: 8px;
+}
+.timeline-label.pipeline .block {
+  background-color: #fff59d !important;
 }
 </style>
